Move code for Accordion-BS-test to BSAccordionTests.cs and update code by using Link control for accessing to <a> in AccordionsPage.cs
