+++ src/Atata.Bootstrap.Tests/Atata.Bootstrap.Tests.csproj
@@ -65,6 +65,7 @@
       <Link>Properties\CommonAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="AtataSettings.cs" />
+    <Compile Include="BSAccordionTests.cs" />
     <Compile Include="Components\AccordionsPage.cs" />
     <Compile Include="UITestFixture.cs" />
     <Compile Include="BSNavTests.cs" />
