+++ src/Atata.Bootstrap.Tests/Components/AccordionsPage.cs
@@ -8,54 +8,22 @@ namespace Atata.Bootstrap.Tests
     {
         public H1<_> Header { get; set; }
 
-        public Text<_> Content { get; set; }
+        public new Text<_> Content { get; set; }
 
-        public BSAccordion<_> Menu1 { get; private set; }
-
-        public BSAccordion<_> Menu2 { get; private set; }
-
-        public BSAccordion<_> Menu3 { get; private set; }
-
-        public BSAccordion<_> Menu4 { get; private set; }
-
+        public Link<_> Menu1 { get; set; }
+        public Link<_> Menu2 { get; set; }
+        public Link<_> Menu3 { get; set; }
 
         public ControlList<AccordionItem, _> AccordionItems { get; private set; }
 
-        [FindByClass("active")]
-        public BSPill<_> ActiveMenu { get; private set; }
-
-        [ControlDefinition("div", ContainingClass = "panel-heading", ComponentTypeName = "accordion item")]
+        [ControlDefinition("div", ContainingClass = "panel-heading")]
         public class AccordionItem : Control<_>
         {
             public H4<_> Header { get; private set; }
 
-            [ControlDefinition("a", ContainingClass = "collapsed", ComponentTypeName = "accordion item")]
+            [ControlDefinition("a", ContainingClass = "collapsed")]
             public Control<_> ATag { get; private set; }
-
-            [FindByXPath("p")]
-            public Text<_> Text { get; private set; }
-
-            [FindFirst]
-            public TextInput<_> TextInput { get; private set; }
         }
 
-        public class TabPane2 : BSTabPane<_>
-        {
-            public H3<_> Header { get; private set; }
-
-            [FindByXPath("p")]
-            public Text<_> Text { get; private set; }
-
-            [FindFirst]
-            public CheckBox<_> CheckBox { get; private set; }
-        }
-
-        public class TabPane3 : BSTabPane<_>
-        {
-            public H3<_> Header { get; private set; }
-
-            [FindByXPath("p")]
-            public Text<_> Text { get; private set; }
-        }
     }
 }
